Project list update (#43)
* prolog app info

insert the prolog app info and replace the `weather` project page with
it. Insert assets, mdx info, and adjust the width of the project page.

* update prolog app with gif exapmle, update type of project info for
instances where the demo link is not provided.

* project space space

adjust the spacing between headings and paragraph elements.

* markdown project

update the markdown project with more infomration, better syntax, and
code blocks to provide better context to the reader.

* fullstack-ent update

update the markdown for fullstack-ent, fixing multiple grammar mistakes,
inserting code-blocks, and providing improved explanations.

* update markdown page with the refactor markdown information

* style and markdown update

adjust spacing for some elements in the project page, removed the image
from the prolog page and removed a duplicate title from the markdown
page.

* project text update

update the resume, remove old project information, and update the
homepage.

* update portfolio gif showcase

* adjust spacin mistake and apply text underline to link in about section

* href error

removed the invalid `a` element, render regular text isntead if there is
no link to a live project.